completed logger : moved ErrorTracking header Logger class methods are now static Added source to CMakeLists moved log file to a 'log' subfolder log file now has .log extension changed getDate to make it more idiomatic and shorter you can't return nullptr instead of a std::string, getType now throws an exception on error
