+++ src/CMakeLists.txt
@@ -3,4 +3,5 @@ cmake_minimum_required(VERSION 3.13.0)
 target_sources(bonjourmonde PRIVATE
     main.cpp
     System.cpp
+    Logger.cpp
 )
