+++ src/Logger.cpp
@@ -5,24 +5,17 @@
 #include <ctime>
 #include <iomanip>
 #include <fstream>
-
-Logger::Logger(/* args */)
-{
-}
-
-Logger::~Logger()
-{
-}
+#include <iomanip>
 
 bool Logger::log(MESSAGE type, std::string message, std::string location)
 {
     std::fstream fs;
+    // create log folder if it doesn't exist
     std::string path = System::resolvePath(std::vector<std::string> {
-        getDate()
+        "log", getDate().append(".log")
     });
-    getDate();
 
-    fs.open(path, std::fstream::in | std::fstream::out | std::fstream::app);
+    fs.open(path, std::fstream::out | std::fstream::app);
 
     fs << getTime() << std::endl;
     fs << getType(type) << ": " << message << std::endl;
@@ -37,18 +30,11 @@ std::string Logger::getDate()
 {
     time_t ttime = time(0);
     tm *local_time = localtime(&ttime);
-    std::string res = "";
-    
-    res.append(std::to_string(local_time->tm_mday));
-    res.append("-");
-    res.append(std::to_string(local_time->tm_mon + 1));
-    res.append("-");
-    res.append(std::to_string(local_time->tm_year + 1900));
-    return(res);
-//    std::cout << "Year: "<< 1900 + local_time->tm_year << std::endl;
-//    std::cout << "Month: "<< 1 + local_time->tm_mon<< std::endl;
-//    std::cout << "Day: "<< local_time->tm_mday << std::endl;
-//    std::cout << "Time: "<< local_time->tm_hour << ":" << local_time->tm_min << ":" << local_time->tm_sec << std::endl;
+    std::stringstream builder;
+
+    builder << std::put_time(local_time, "%d-%m-%Y");
+
+    return(builder.str());
 }
 
 std::string Logger::getTime()
@@ -73,6 +59,6 @@ std::string Logger::getType(MESSAGE type)
         return("WARNING");
     else if (type == ERROR)
         return("ERROR");
-    else
-        return(nullptr);
+    
+    throw std::runtime_error("wrong MESSAGE value");
 }
